Start refactor of certificate validation.
Break apart certificate validation from a single compilation target
into multiple ones.  This will allow us to invoke Globus from the
credential fetcher process but VOMS from either parent or child.

Also moves the validation function wholesale to the credential
fetcher process.

Doesn't actually implement serialization and movement of the VOMS
data object from child back to parent.
