+++ cvmfs/voms_authz/voms_authz.cc
@@ -20,167 +20,13 @@
 #include <vector>
 
 #include "fuse/fuse_lowlevel.h"
-#include "voms/voms_apic.h"
-
-#include "globus/globus_gsi_cert_utils.h"
-#include "globus/globus_gsi_credential.h"
-#include "globus/globus_module.h"
 
 #include "../logging.h"
 #include "../platform.h"
 #include "../util.h"
+#include "voms_lib.h"
 
 // TODO(jblomer): add unit tests for static functions
-
-extern "C" {
-// VOMS API declarations
-struct vomsdata * (*g_VOMS_Init)(char *voms, char *cert) = NULL;
-void (*g_VOMS_Destroy)(struct vomsdata *vd) = NULL;
-int (*g_VOMS_Retrieve)(X509 *cert, STACK_OF(X509) *chain, int how,
-                         struct vomsdata *vd, int *error) = NULL;
-char * (*g_VOMS_ErrorMessage)(struct vomsdata *vd, int error, char *buffer,
-                              int len) = NULL;
-
-// Globus API declarations
-
-// For activating / deactivating modules.
-int (*g_globus_module_activate)(
-    globus_module_descriptor_t * module_descriptor) = NULL;
-int (*g_globus_module_deactivate)(
-    globus_module_descriptor_t * module_descriptor) = NULL;
-int (*g_globus_thread_set_model)(
-    const char *                        model) = NULL;
-globus_object_t * (*g_globus_error_get)(
-    globus_result_t                     result) = NULL;
-char * (*g_globus_error_print_chain)(
-    globus_object_t *                   error) = NULL;
-globus_object_t *g_GLOBUS_ERROR_BASE_STATIC_PROTOTYPE = NULL;
-// Modules we'll want to use.
-globus_module_descriptor_t *g_globus_i_gsi_cert_utils_module = NULL;
-globus_module_descriptor_t *g_globus_i_gsi_credential_module = NULL;
-globus_module_descriptor_t *g_globus_i_gsi_callback_module = NULL;
-globus_module_descriptor_t *g_globus_i_gsi_sysconfig_module = NULL;
-
-// Actual functions we need to invoke.
-globus_result_t (*g_globus_gsi_cred_handle_init)(
-    globus_gsi_cred_handle_t *          handle,
-    globus_gsi_cred_handle_attrs_t      handle_attrs) = NULL;
-globus_result_t (*g_globus_gsi_cred_handle_destroy)(
-    globus_gsi_cred_handle_t            handle) = NULL;
-globus_result_t (*g_globus_gsi_cred_read_proxy_bio)(
-    globus_gsi_cred_handle_t            handle,
-    BIO *                               bio) = NULL;
-globus_result_t (*g_globus_gsi_cred_get_cert)(
-    globus_gsi_cred_handle_t            handle,
-    X509 **                             cert) = NULL;
-globus_result_t (*g_globus_gsi_cred_get_key)(
-    globus_gsi_cred_handle_t            handle,
-    EVP_PKEY **                         key) = NULL;
-globus_result_t (*g_globus_gsi_cred_get_cert_chain)(
-    globus_gsi_cred_handle_t            handle,
-    STACK_OF(X509) **                   cert_chain) = NULL;
-globus_result_t (*g_globus_gsi_cred_get_subject_name)(
-    globus_gsi_cred_handle_t            handle,
-    char **                             subject_name) = NULL;
-globus_result_t (*g_globus_gsi_cred_verify_cert_chain)(
-    globus_gsi_cred_handle_t            cred_handle,
-    globus_gsi_callback_data_t          callback_data) = NULL;
-globus_result_t (*g_globus_gsi_cert_utils_get_cert_type)(
-    X509 *                              cert,
-    globus_gsi_cert_utils_cert_type_t * type) = NULL;
-globus_result_t (*g_globus_gsi_cert_utils_get_identity_cert)(
-    STACK_OF(X509) *                    cert_chain,
-    X509 **                             eec) = NULL;
-globus_result_t (*g_globus_gsi_callback_data_init)(
-    globus_gsi_callback_data_t *        callback_data) = NULL;
-globus_result_t (*g_globus_gsi_callback_data_destroy)(
-    globus_gsi_callback_data_t          callback_data) = NULL;
-globus_result_t (*g_globus_gsi_callback_set_cert_dir)(
-    globus_gsi_callback_data_t          callback_data,
-    char *                              cert_dir) = NULL;
-globus_result_t (*g_globus_gsi_sysconfig_get_cert_dir_unix)(
-    char **                             cert_dir) = NULL;
-
-
-
-// Although it is OK to use strcmp for 99% of DNs, there are some string rules
-// in X509 about handling of multiple spaces ("  "); this will take care of
-// the last 1% of cases.
-int (*g_globus_i_gsi_cert_utils_dn_cmp)(
-    const char *                        dn1,
-    const char *                        dn2) = NULL;
-bool g_globus_ok = false;  // Set to true if all globus initialization worked.
-}
-
-
-static bool
-OpenDynLib(void **handle, const char *name, const char *friendly_name) {
-  if (!handle) {return false;}
-  *handle = dlopen(name, RTLD_LAZY);
-  if (!(*handle)) {
-    LogCvmfs(kLogVoms, kLogDebug, "Failed to load %s library; "
-             "authz will not be available.  %s", friendly_name, dlerror());
-  }
-  return *handle;
-}
-
-
-static void
-CloseDynLib(void **handle, const char *name) {
-  if (!handle) {return;}
-  if (*handle && dlclose(*handle)) {
-    LogCvmfs(kLogVoms, kLogDebug, "Failed to unload %s library: %s",
-                                  name, dlerror());
-  }
-  *handle = NULL;
-}
-
-
-template<typename T>
-bool LoadSymbol(void *lib_handle, T *sym, const char *name) {
-  if (!sym) {return false;}
-  *sym = reinterpret_cast<typeof(T)>(dlsym(lib_handle, name));
-  if (!(*sym)) {
-    LogCvmfs(kLogVoms, kLogDebug, "Failed to load %s: %s", name, dlerror());
-  }
-  return *sym;
-}
-
-
-static void
-PrintGlobusError(globus_result_t result) {
-  globus_object_t *error_obj = (*g_globus_error_get)(result);
-  if (error_obj == g_GLOBUS_ERROR_BASE_STATIC_PROTOTYPE) {
-    LogCvmfs(kLogVoms, kLogDebug, "Globus error occurred (no further "
-                                  "information available)");
-    return;
-  }
-  char *error_full = (*g_globus_error_print_chain)(error_obj);
-  if (!error_full) {
-    LogCvmfs(kLogVoms, kLogDebug, "Globus error occurred (error unprintable)");
-    return;
-  }
-  LogCvmfs(kLogVoms, kLogDebug, "Globus error: %s", error_full);
-  free(error_full);
-}
-
-
-struct authz_data {
-  struct vomsdata *voms_;
-  char *dn_;
-
-  authz_data() :
-    voms_(NULL),
-    dn_(NULL)
-  {}
-
-  ~authz_data() {
-    if (voms_) {(*g_VOMS_Destroy)(voms_);}
-    if (dn_) {OPENSSL_free(dn_);}
-  }
-};
-
-
 // TODO(jblomer): add comment
 // TODO(jblomer): unit test the class
 // TODO(jblomer): Member functions should start with a capital letter
@@ -189,21 +35,9 @@ struct authz_data {
 // TODO(jblomer): member naming: trailing underscore, no m_ prefix
 class AuthzSessionCache {
   AuthzSessionCache()
-    : m_zombie(true),
-      m_last_clean(platform_monotonic_time()),
-      m_globus_module_handle(NULL),
-      m_globus_gsi_cert_utils_handle(NULL),
-      m_globus_gsi_credential_handle(NULL),
-      m_globus_gsi_callback_handle(NULL)
+    : m_last_clean(platform_monotonic_time())
   {
     pthread_mutex_init(&m_mutex, NULL);
-    load_voms_library();
-    if (!m_zombie) {
-      m_zombie = true;
-      load_globus_library();
-    }
-    LogCvmfs(kLogVoms, kLogDebug|kLogSyslog, "Support for authz is %senabled.",
-      m_zombie ? "NOT " : "");
   }
 
   ~AuthzSessionCache() {
@@ -213,14 +47,11 @@ class AuthzSessionCache {
     {
       delete (it->second.first);
     }
-    close_voms_library();
-    close_globus_library();
     pthread_mutex_destroy(&m_mutex);
   }
 
  public:
   bool get(pid_t pid, authz_data **authz) {  // NOLINT
-    if (m_zombie) {return false;}
     time_t now = platform_monotonic_time();
     KeyType mykey;
     if (!lookup(pid, mykey)) {return false;}
@@ -253,7 +84,6 @@ class AuthzSessionCache {
    */
   authz_data *try_put(pid_t pid, authz_data *authz_ptr)
   {
-    if (m_zombie) {return NULL;}
     KeyType mykey;
     if (!lookup(pid, mykey)) {
       LogCvmfs(kLogVoms, kLogDebug, "Failed to determine session key "
@@ -413,403 +243,17 @@ class AuthzSessionCache {
     }
   }
 
-  void load_voms_library() {
-    if (!OpenDynLib(&m_libvoms_handle, "libvomsapi.so.1", "VOMS")) {return;}
-    if (
-       !LoadSymbol(m_libvoms_handle, &g_VOMS_Init, "VOMS_Init") ||
-       !LoadSymbol(m_libvoms_handle, &g_VOMS_Destroy, "VOMS_Destroy") ||
-       !LoadSymbol(m_libvoms_handle, &g_VOMS_Retrieve, "VOMS_Retrieve") ||
-       !LoadSymbol(m_libvoms_handle, &g_VOMS_Destroy, "VOMS_Destroy") ||
-       !LoadSymbol(m_libvoms_handle, &g_VOMS_ErrorMessage, "VOMS_ErrorMessage")
-      ) {
-        g_VOMS_Init = NULL;
-        return;
-    }
-    LogCvmfs(kLogVoms, kLogDebug, "Successfully loaded VOMS library");
-    m_zombie = false;
-  }
 
-
-  void load_globus_library() {
-    if (!OpenDynLib(&m_globus_module_handle, "libglobus_common.so.0",
-                    "Globus common")) {return;}
-    if (!OpenDynLib(&m_globus_gsi_cert_utils_handle,
-                    "libglobus_gsi_cert_utils.so.0",
-                    "Globus GSI cert utils")) {return;}
-    if (!OpenDynLib(&m_globus_gsi_credential_handle,
-                    "libglobus_gsi_credential.so.1",
-                    "Globus GSI credential")) {return;}
-    if (!OpenDynLib(&m_globus_gsi_callback_handle,
-                    "libglobus_gsi_callback.so.0",
-                    "Globus GSI callback")) {return;}
-    if (!OpenDynLib(&m_globus_gsi_sysconfig_handle,
-                    "libglobus_gsi_sysconfig.so.1",
-                    "Globus GSI sysconfig")) {return;}
-    if (
-       !LoadSymbol(m_globus_module_handle, &g_globus_module_activate,
-                   "globus_module_activate") ||
-       !LoadSymbol(m_globus_module_handle, &g_globus_module_deactivate,
-                   "globus_module_deactivate") ||
-       !LoadSymbol(m_globus_module_handle, &g_globus_thread_set_model,
-                   "globus_thread_set_model") ||
-       !LoadSymbol(m_globus_module_handle, &g_globus_error_get,
-                   "globus_error_get") ||
-       !LoadSymbol(m_globus_module_handle, &g_globus_error_print_chain,
-                   "globus_error_print_chain") ||
-       !LoadSymbol(m_globus_module_handle,
-                   &g_GLOBUS_ERROR_BASE_STATIC_PROTOTYPE,
-                   "GLOBUS_ERROR_BASE_STATIC_PROTOTYPE") ||
-       !LoadSymbol(m_globus_gsi_cert_utils_handle,
-                   &g_globus_i_gsi_cert_utils_module,
-                   "globus_i_gsi_cert_utils_module") ||
-       !LoadSymbol(m_globus_gsi_credential_handle,
-                   &g_globus_i_gsi_credential_module,
-                   "globus_i_gsi_credential_module") ||
-       !LoadSymbol(m_globus_gsi_callback_handle,
-                   &g_globus_i_gsi_callback_module,
-                   "globus_i_gsi_callback_module") ||
-       !LoadSymbol(m_globus_gsi_sysconfig_handle,
-                   &g_globus_i_gsi_sysconfig_module,
-                   "globus_i_gsi_sysconfig_module") ||
-       !LoadSymbol(m_globus_gsi_credential_handle,
-                   &g_globus_gsi_cred_handle_init,
-                   "globus_gsi_cred_handle_init") ||
-       !LoadSymbol(m_globus_gsi_credential_handle,
-                   &g_globus_gsi_cred_handle_destroy,
-                   "globus_gsi_cred_handle_destroy") ||
-       !LoadSymbol(m_globus_gsi_credential_handle,
-                   &g_globus_gsi_cred_read_proxy_bio,
-                   "globus_gsi_cred_read_proxy_bio") ||
-       !LoadSymbol(m_globus_gsi_credential_handle,
-                   &g_globus_gsi_cred_verify_cert_chain,
-                   "globus_gsi_cred_verify_cert_chain") ||
-       !LoadSymbol(m_globus_gsi_credential_handle,
-                   &g_globus_gsi_cred_get_cert,
-                   "globus_gsi_cred_get_cert") ||
-       !LoadSymbol(m_globus_gsi_credential_handle,
-                   &g_globus_gsi_cred_get_key,
-                   "globus_gsi_cred_get_key") ||
-       !LoadSymbol(m_globus_gsi_credential_handle,
-                   &g_globus_gsi_cred_get_cert_chain,
-                   "globus_gsi_cred_get_cert_chain") ||
-       !LoadSymbol(m_globus_gsi_credential_handle,
-                   &g_globus_gsi_cred_get_subject_name,
-                   "globus_gsi_cred_get_subject_name") ||
-       !LoadSymbol(m_globus_gsi_cert_utils_handle,
-                   &g_globus_gsi_cert_utils_get_cert_type,
-                   "globus_gsi_cert_utils_get_cert_type") ||
-       !LoadSymbol(m_globus_gsi_cert_utils_handle,
-                   &g_globus_i_gsi_cert_utils_dn_cmp,
-                   "globus_i_gsi_cert_utils_dn_cmp") ||
-       !LoadSymbol(m_globus_gsi_cert_utils_handle,
-                   &g_globus_gsi_cert_utils_get_identity_cert,
-                   "globus_gsi_cert_utils_get_identity_cert") ||
-       !LoadSymbol(m_globus_gsi_callback_handle,
-                   &g_globus_gsi_callback_data_init,
-                   "globus_gsi_callback_data_init") ||
-       !LoadSymbol(m_globus_gsi_callback_handle,
-                   &g_globus_gsi_callback_data_destroy,
-                   "globus_gsi_callback_data_destroy") ||
-       !LoadSymbol(m_globus_gsi_callback_handle,
-                   &g_globus_gsi_callback_set_cert_dir,
-                   "globus_gsi_callback_set_cert_dir") ||
-       !LoadSymbol(m_globus_gsi_sysconfig_handle,
-                   &g_globus_gsi_sysconfig_get_cert_dir_unix,
-                   "globus_gsi_sysconfig_get_cert_dir_unix")
-      ) {return;}
-    if (GLOBUS_SUCCESS !=
-        (g_globus_thread_set_model)("none")) {
-      LogCvmfs(kLogVoms, kLogDebug, "Failed to enable Globus thread model.");
-      return;
-    }
-    if (GLOBUS_SUCCESS !=
-        (*g_globus_module_activate)(g_globus_i_gsi_cert_utils_module)) {
-      LogCvmfs(kLogVoms, kLogDebug, "Failed to activate Globus GSI cert utils"
-                                    " module.");
-      return;
-    }
-    if (GLOBUS_SUCCESS !=
-        (*g_globus_module_activate)(g_globus_i_gsi_credential_module)) {
-      LogCvmfs(kLogVoms, kLogDebug, "Failed to activate Globus GSI credential"
-                                    " module.");
-      (*g_globus_module_deactivate)(g_globus_i_gsi_cert_utils_module);
-      return;
-    }
-    if (GLOBUS_SUCCESS !=
-        (*g_globus_module_activate)(g_globus_i_gsi_callback_module)) {
-      (*g_globus_module_deactivate)(g_globus_i_gsi_cert_utils_module);
-      (*g_globus_module_deactivate)(g_globus_i_gsi_credential_module);
-      return;
-    }
-    if (GLOBUS_SUCCESS !=
-        (*g_globus_module_activate)(g_globus_i_gsi_sysconfig_module)) {
-      (*g_globus_module_deactivate)(g_globus_i_gsi_cert_utils_module);
-      (*g_globus_module_deactivate)(g_globus_i_gsi_credential_module);
-      (*g_globus_module_deactivate)(g_globus_i_gsi_callback_module);
-    }
-    LogCvmfs(kLogVoms, kLogDebug, "Successfully loaded Globus library");
-    m_zombie = false;
-    g_globus_ok = true;
-  }
-
-
-  void close_voms_library() {
-    CloseDynLib(&m_libvoms_handle, "VOMS");
-    g_VOMS_Init = NULL;
-    g_VOMS_Destroy = NULL;
-    g_VOMS_Retrieve = NULL;
-    g_VOMS_ErrorMessage = NULL;
-  }
-
-
-  void close_globus_library() {
-    if (g_globus_ok) {
-      (*g_globus_module_deactivate)(g_globus_i_gsi_cert_utils_module);
-      (*g_globus_module_deactivate)(g_globus_i_gsi_credential_module);
-      (*g_globus_module_deactivate)(g_globus_i_gsi_callback_module);
-      (*g_globus_module_deactivate)(g_globus_i_gsi_sysconfig_module);
-    }
-    CloseDynLib(&m_globus_module_handle, "Globus module");
-    CloseDynLib(&m_globus_gsi_cert_utils_handle, "Globus GSI cert utils");
-    CloseDynLib(&m_globus_gsi_credential_handle, "Globus GSI credential");
-    CloseDynLib(&m_globus_gsi_callback_handle, "Globus GSI callback");
-    CloseDynLib(&m_globus_gsi_sysconfig_handle, "Globus GSI sysconfig");
-    g_globus_thread_set_model = NULL;
-    g_globus_module_activate = NULL;
-    g_globus_module_deactivate = NULL;
-    g_globus_error_get = NULL;
-    g_globus_error_print_chain = NULL;
-    g_GLOBUS_ERROR_BASE_STATIC_PROTOTYPE = NULL;
-    g_globus_i_gsi_cert_utils_module = NULL;
-    g_globus_i_gsi_credential_module = NULL;
-    g_globus_i_gsi_callback_module = NULL;
-    g_globus_i_gsi_sysconfig_module = NULL;
-    g_globus_gsi_cred_handle_init = NULL;
-    g_globus_gsi_cred_handle_destroy = NULL;
-    g_globus_gsi_cred_get_key = NULL;
-    g_globus_gsi_cred_verify_cert_chain = NULL;
-    g_globus_gsi_cred_get_subject_name = NULL;
-    g_globus_gsi_cert_utils_get_cert_type = NULL;
-    g_globus_gsi_cert_utils_get_identity_cert = NULL;
-    g_globus_i_gsi_cert_utils_dn_cmp = NULL;
-    g_globus_gsi_callback_data_init = NULL;
-    g_globus_gsi_callback_data_destroy = NULL;
-    g_globus_gsi_callback_set_cert_dir = NULL;
-    g_globus_gsi_sysconfig_get_cert_dir_unix = NULL;
-    g_globus_ok = false;
-  }
-
-
-  bool m_zombie;
   time_t m_last_clean;
   pthread_mutex_t m_mutex;
 
   KeyToVOMS m_map;
-
-  // Various library handles.
-  void *m_libvoms_handle;
-  void *m_globus_module_handle;
-  void *m_globus_gsi_cert_utils_handle;
-  void *m_globus_gsi_credential_handle;
-  void *m_globus_gsi_callback_handle;
-  void *m_globus_gsi_sysconfig_handle;
 
   static AuthzSessionCache g_cache;
 };
 
 
 AuthzSessionCache AuthzSessionCache::g_cache;
-
-
-static FILE *
-GetProxyFile(const struct fuse_ctx *ctx) {
-  return GetProxyFile(ctx->pid, ctx->uid, ctx->gid);
-}
-
-
-/**
- * Create the VOMS data structure to the best of our abilities.
- *
- * Resulting memory is owned by caller and must be destroyed with VOMS_Destroy.
- */
-struct authz_state {
-  FILE *m_fp;
-  struct vomsdata *m_voms;
-  globus_gsi_cred_handle_t m_cred;
-  BIO *m_bio;
-  X509 *m_cert;
-  EVP_PKEY *m_pkey;
-  STACK_OF(X509) *m_chain;
-  char *m_subject;
-  globus_gsi_callback_data_t m_callback;
-
-  authz_state() :
-    m_fp(NULL),
-    m_voms(NULL),
-    m_cred(NULL),
-    m_bio(NULL),
-    m_cert(NULL),
-    m_pkey(NULL),
-    m_chain(NULL),
-    m_subject(NULL),
-    m_callback(NULL)
-  {}
-
-  ~authz_state() {
-    if (m_fp) {fclose(m_fp);}
-    if (m_voms) {(*g_VOMS_Destroy)(m_voms);}
-    if (m_cred) {(*g_globus_gsi_cred_handle_destroy)(m_cred);}
-    if (m_bio) {BIO_free(m_bio);}
-    if (m_cert) {X509_free(m_cert);}
-    if (m_pkey) {EVP_PKEY_free(m_pkey);}
-    if (m_chain) {sk_X509_pop_free(m_chain, X509_free);}
-    if (m_subject) {OPENSSL_free(m_subject);}
-    if (m_callback) {(*g_globus_gsi_callback_data_destroy)(m_callback);}
-  }
-};
-
-
-static authz_data*
-GenerateVOMSData(const struct fuse_ctx *ctx)
-{
-  authz_state state;
-
-  state.m_fp = GetProxyFile(ctx);
-  if (!state.m_fp) {
-    LogCvmfs(kLogVoms, kLogDebug, "Could not find process's proxy file.");
-    return NULL;
-  }
-
-  // Start of Globus proxy parsing and verification...
-  globus_result_t result =
-      (*g_globus_gsi_cred_handle_init)(&state.m_cred, NULL);
-  if (GLOBUS_SUCCESS != result) {
-    PrintGlobusError(result);
-    return NULL;
-  }
-
-  state.m_bio = BIO_new_fp(state.m_fp, 0);
-  if (!state.m_bio) {
-    LogCvmfs(kLogVoms, kLogDebug, "Unable to allocate new BIO object");
-    return NULL;
-  }
-
-  result = (*g_globus_gsi_cred_read_proxy_bio)(state.m_cred, state.m_bio);
-  if (GLOBUS_SUCCESS != result) {
-    LogCvmfs(kLogVoms, kLogDebug, "Failed to parse credentials");
-    PrintGlobusError(result);
-    return NULL;
-  }
-
-  // Setup Globus callback object.
-  result = (*g_globus_gsi_callback_data_init)(&state.m_callback);
-  if (GLOBUS_SUCCESS != result) {
-    PrintGlobusError(result);
-    return NULL;
-  }
-  char *cert_dir;
-  result = (*g_globus_gsi_sysconfig_get_cert_dir_unix)(&cert_dir);
-  if (GLOBUS_SUCCESS != result) {
-    LogCvmfs(kLogVoms, kLogDebug, "Failed to determine trusted certificates "
-                                  "directory.");
-    PrintGlobusError(result);
-    return NULL;
-  }
-  result = (*g_globus_gsi_callback_set_cert_dir)(state.m_callback, cert_dir);
-  free(cert_dir);
-  if (GLOBUS_SUCCESS != result) {
-    PrintGlobusError(result);
-    return NULL;
-  }
-
-  // Verify credential chain.
-  result = (*g_globus_gsi_cred_verify_cert_chain)(state.m_cred,
-                                                  state.m_callback);
-  if (GLOBUS_SUCCESS != result) {
-    LogCvmfs(kLogVoms, kLogDebug, "Failed to validate credentials");
-    PrintGlobusError(result);
-    return NULL;
-  }
-
-  // Load key and certificate from Globus handle
-  result = (*g_globus_gsi_cred_get_key)(state.m_cred, &state.m_pkey);
-  if (GLOBUS_SUCCESS != result) {
-    LogCvmfs(kLogVoms, kLogDebug, "Failed to get process private key.");
-    PrintGlobusError(result);
-    return NULL;
-  }
-  result = (*g_globus_gsi_cred_get_cert)(state.m_cred, &state.m_cert);
-  if (GLOBUS_SUCCESS != result) {
-    LogCvmfs(kLogVoms, kLogDebug, "Failed to get process certificate.");
-    PrintGlobusError(result);
-    return NULL;
-  }
-
-  // Check proxy public key and private key match.
-  if (!X509_check_private_key(state.m_cert, state.m_pkey)) {
-    LogCvmfs(kLogVoms, kLogDebug, "Process certificate and key do not match");
-    return NULL;
-  }
-
-  // Load certificate chain
-  result = (*g_globus_gsi_cred_get_cert_chain)(state.m_cred, &state.m_chain);
-  if (GLOBUS_SUCCESS != result) {
-    LogCvmfs(kLogVoms, kLogDebug, "Process does not have cert chain.");
-    PrintGlobusError(result);
-    return NULL;
-  }
-
-  // Look through certificates to find an EEC (which has the subject)
-  globus_gsi_cert_utils_cert_type_t cert_type;
-  X509 *eec_cert = state.m_cert;
-  result = (*g_globus_gsi_cert_utils_get_cert_type)(state.m_cert, &cert_type);
-  if (GLOBUS_SUCCESS != result) {
-    PrintGlobusError(result);
-    return NULL;
-  }
-  if (!(cert_type & GLOBUS_GSI_CERT_UTILS_TYPE_EEC)) {
-    result = (*g_globus_gsi_cert_utils_get_identity_cert)(state.m_chain,
-                                                          &eec_cert);
-    if (GLOBUS_SUCCESS != result) {
-      PrintGlobusError(result);
-      return NULL;
-    }
-  }
-  // From the EEC, use OpenSSL to determine the subject
-  state.m_subject = X509_NAME_oneline(X509_get_subject_name(eec_cert), NULL, 0);
-  if (!state.m_subject) {
-    LogCvmfs(kLogVoms, kLogDebug, "Unable to determine certificate DN.");
-    return NULL;
-  }
-
-  state.m_voms = (*g_VOMS_Init)(NULL, NULL);
-  if (!state.m_voms) {
-    return NULL;
-  }
-
-  int voms_error = 0;
-  const int retval = (*g_VOMS_Retrieve)(state.m_cert, state.m_chain,
-                                        RECURSE_CHAIN,
-                                        state.m_voms, &voms_error);
-  // If there is no VOMS extension (VERR_NOEXT), this shouldn't be fatal.
-  if (!retval && (voms_error != VERR_NOEXT)) {
-    char *err_str = (*g_VOMS_ErrorMessage)(state.m_voms, voms_error, NULL, 0);
-    LogCvmfs(kLogVoms, kLogDebug, "Unable to parse VOMS file: %s\n",
-             err_str);
-    free(err_str);
-    return NULL;
-  }
-
-  // Move pointers to returned authz_data structure
-  authz_data *authz = new authz_data();
-  authz->dn_ = state.m_subject;
-  state.m_subject = NULL;
-  if (voms_error != VERR_NOEXT) {
-    authz->voms_ = state.m_voms;
-    state.m_voms = NULL;
-  }
-  return authz;
-}
 
 
 // TODO(jblomer): can probably be replaced by SplitString from util
@@ -869,7 +313,7 @@ static bool CheckMultipleAuthz(const authz_data *authz_ptr,
 
 
 bool CheckVOMSAuthz(const struct fuse_ctx *ctx, const std::string & authz) {
-  if (g_VOMS_Init == NULL || !g_globus_ok) {
+  if (!VomsLib::GetInstance().IsValid()) {
     LogCvmfs(kLogVoms, kLogSyslog | kLogDebug,
              "VOMS library not present; failing VOMS authz.");
     return false;
@@ -881,7 +325,7 @@ bool CheckVOMSAuthz(const struct fuse_ctx *ctx, const std::string & authz) {
   // Get VOMS information from cache; if not present, store it.
   authz_data *authz_ptr;
   if (!AuthzSessionCache::GetInstance().get(ctx->pid, &authz_ptr)) {
-    authz_ptr = GenerateVOMSData(ctx);
+    authz_ptr = GetAuthzData(ctx->pid, ctx->uid, ctx->gid);
     if (authz_ptr) {
       authz_ptr = AuthzSessionCache::GetInstance().try_put(ctx->pid, authz_ptr);
       LogCvmfs(kLogVoms, kLogDebug,
@@ -960,7 +404,7 @@ CheckSingleAuthz(const authz_data *authz_ptr, const std::string & authz)
   // Now we have valid data, check authz.
   // First, check if it is a valid DN.
   if (is_dn) {
-    return !(*g_globus_i_gsi_cert_utils_dn_cmp)(authz.c_str(), authz_ptr->dn_);
+    return !strcmp(authz.c_str(), authz_ptr->dn_);
   }
   // If there is no VOMS info, return immediately..
   if (!authz_ptr->voms_) {return false;}
