+++ cvmfs/voms_authz/voms_authz.h
@@ -13,6 +13,7 @@
 #include <string>
 
 struct fuse_ctx;
+struct authz_data;
 
 bool CheckVOMSAuthz(const struct fuse_ctx *ctx, const std::string &);
 
@@ -21,6 +22,7 @@ bool ConfigureCurlHandle(void *curl_handle, pid_t pid, uid_t uid, gid_t gid,
 void ReleaseCurlHandle(void *curl_handle, void *data);
 
 FILE *GetProxyFile(pid_t pid, uid_t uid, gid_t gid);
+authz_data* GetAuthzData(pid_t pid, uid_t uid, gid_t gid);
 
 #endif  // CVMFS_VOMS_AUTHZ_VOMS_AUTHZ_H_
