+++ test/unittests/t_voms.cc
@@ -7,6 +7,8 @@
 // Single Authz test is static
 // TODO(jblomer): find a better solution than pulling in the .cc file
 #include "../../cvmfs/voms_authz/voms_authz.cc"  // NOLINT
+#include "../../cvmfs/voms_authz/voms_lib.cc"  // NOLINT
+#include "../../cvmfs/voms_authz/loader_helper.cc"  // NOLINT
 
 #define TEST_DN "/DC=ch/DC=cern/OU=Organic Units/OU=Users" \
                 "/CN=bbockelm/CN=659869/CN=Brian Paul Bockelman"
